requirements.txt
# dependencies
beautifulsoup4==4.12.2
colorama==0.4.6
duckduckgo_search==3.9.8
openai~=1.3.3
playwright==1.39.0
python-dotenv~=1.0.0
pyyaml==6.0.1
selenium==4.15.2
webdriver-manager==4.0.1
pydantic==2.5.1
python-multipart==0.0.6
markdown==3.5.1
langchain==0.0.335
tavily-python==0.2.8
permchain==0.0.6
arxiv==2.0.0
PyMuPDF==1.23.6
requests==2.31.0
jinja2==3.1.2
loguru
markdown
tenacity
tiktoken
faiss_cpu